refactored initial classes into files
