+++ player.rb
@@ -1,0 +1,14 @@
+class Player
+  WHITE = 'w'
+  BLACK = 'b'
+
+  attr_reader :color
+
+  def initialize(params)
+    @color  = params.delete(:color)
+  end
+
+  private
+
+end
+
